+++ straxen/plugins/__init__.py
@@ -59,9 +59,3 @@ from .acqmon_processing import *
 
 from . import online_monitor
 from .online_monitor import *
-
-from . import peak_shadow
-from .peak_shadow import *
-
-from . import event_shadow
-from .event_shadow import *
